requirements.txt
google-generativeai
feedparser
requests
trafilatura
python-dotenv